simple-frontend/requirements.txt
gradio
requests
Pillow
dotenv